requirements.txt
torch
transformers
sentence-transformers
faiss-cpu
numpy
pandas
tqdm
scikit-learn
fastapi
uvicorn